refactor: extraction all from common folder
