+++ src/sc_kpm/__init__.py
@@ -7,7 +7,7 @@ Distributed under the MIT License
 from sc_client import models as sc_models
 from sc_client.constants import sc_types
 
-from sc_kpm.common.sc_agent import ScAgent
-from sc_kpm.common.sc_keynodes import ScKeynodes
-from sc_kpm.common.sc_module import ScModule
-from sc_kpm.common.sc_server import ScServer
+from sc_kpm.sc_agent import ScAgent
+from sc_kpm.sc_keynodes import ScKeynodes
+from sc_kpm.sc_module import ScModule
+from sc_kpm.sc_server import ScServer
