--- src/sc_kpm/common/__init__.py
@@ -1,7 +1,0 @@
-"""
-This source file is part of an OSTIS project. For the latest info, see https://github.com/ostis-ai
-Distributed under the MIT License
-(See an accompanying file LICENSE or a copy at https://opensource.org/licenses/MIT)
-"""
-
-from sc_kpm.common.identifiers import CommonIdentifiers, QuestionStatus, ScAlias
