+++ src/sc_kpm/utils/action_utils.py
@@ -13,8 +13,7 @@ from sc_client.constants import sc_types
 from sc_client.models import ScAddr, ScConstruction, ScTemplate
 
 from sc_kpm import ScKeynodes
-from sc_kpm.common import CommonIdentifiers, QuestionStatus, ScAlias
-from sc_kpm.common.identifiers import Idtf
+from sc_kpm.identifiers import CommonIdentifiers, Idtf, QuestionStatus, ScAlias
 from sc_kpm.utils.common_utils import check_edge, create_edge, create_node, create_role_relation
 from sc_kpm.utils.retrieve_utils import _get_first_search_template_result
