+++ src/sc_kpm/utils/creation_utils.py
@@ -9,7 +9,7 @@ from sc_client.constants.sc_types import ScType
 from sc_client.models import ScAddr
 
 from sc_kpm import ScKeynodes
-from sc_kpm.common import CommonIdentifiers
+from sc_kpm.identifiers import CommonIdentifiers
 from sc_kpm.utils.common_utils import create_edge, create_node, create_norole_relation, create_role_relation
