+++ tests/test_utils/test_action_utils.py
@@ -11,7 +11,7 @@ from sc_client.constants.common import ScEventType
 
 from sc_kpm import ScAgent, ScKeynodes, ScModule
 from sc_kpm.common import CommonIdentifiers, QuestionStatus
-from sc_kpm.common.sc_result import ScResult
+from sc_kpm.sc_result import ScResult
 from sc_kpm.utils.action_utils import check_action_class, execute_agent
 from sc_kpm.utils.common_utils import create_edge, create_node, delete_elements
 from tests.common_tests import BaseTestCase, server
